requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
rich==13.8.1
httpx==0.27.2

streamlit==1.37.1
matplotlib==3.9.2

pandas==2.2.2
numpy==1.26.4
duckdb==1.1.2
scikit-learn==1.5.2

sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
qdrant-client==1.9.1

pypdf==4.3.1
markdown-it-py==3.0.0
tiktoken==0.7.0
nltk==3.9.1

pytest==8.3.3

torch==2.3.1
